Updated 'next_number' API to work with unsigned numbers

+++ src/utils.h
@@ -43,10 +43,10 @@ bool char_is_hex(char ch);
 char nibble_to_char(uint8_t nibble);
 
 const char *string_skip_whitespace(const char *str);
-const char *string_next_number(const char *str, int *value);
+const char *string_next_number(const char *str, uint16_t *value);
 const char *string_next_token(const char *str, int *length);
-const char *string_next_hex_number(const char *str, int *value);
-const char *string_next_decimal_number(const char *str, int *value);
+const char *string_next_hex_number(const char *str, uint16_t *value);
+const char *string_next_decimal_number(const char *str, uint16_t *value);
 
 const char *string_to_buffer(const char *str,
                              uint8_t bufferLength, uint8_t *buffer, uint8_t *bufferFilled);
